Fixes ssr bugs

+++ src/app/service/storage/storage.service.ts
@@ -1,7 +1,7 @@
 import { isPlatformServer } from '@angular/common';
 import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { AbstractQuestionGroup } from 'arsnova-click-v2-types/src/questions/questiongroup_abstract';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { DB_NAME, DB_TABLE, STORAGE_KEY } from '../../shared/enums';
 import { IndexedDbService } from './indexed.db.service';
 
@@ -25,6 +25,10 @@ export class StorageService {
   }
 
   public create(table: DB_TABLE, key: string | STORAGE_KEY, value: any): Observable<any> {
+    if (isPlatformServer(this.platformId)) {
+      return of(null);
+    }
+
     if (value instanceof AbstractQuestionGroup) {
       value = value.serialize();
     }
@@ -36,20 +40,36 @@ export class StorageService {
   }
 
   public read(table: DB_TABLE, key: string | STORAGE_KEY): Observable<any> {
+    if (isPlatformServer(this.platformId)) {
+      return of(null);
+    }
+
     return this.indexedDbService.get(table, this.formatKey(key));
   }
 
   public delete(table: DB_TABLE, key: string | STORAGE_KEY): Observable<any> {
+    if (isPlatformServer(this.platformId)) {
+      return of(null);
+    }
+
     return this.indexedDbService.remove(table, this.formatKey(key));
   }
 
   public async getAllQuiznames(): Promise<Array<string>> {
+    if (isPlatformServer(this.platformId)) {
+      return of(null).toPromise();
+    }
+
     return (
       await this.indexedDbService.all(DB_TABLE.QUIZ).toPromise()
     ).map(value => value.id);
   }
 
   public getAll(table: DB_TABLE): Observable<any> {
+    if (isPlatformServer(this.platformId)) {
+      return of(null);
+    }
+
     return this.indexedDbService.all(table);
   }
